Round world, Target Line, Better AI, Safe Mitosis
Firstly, the world is round. So cells going up will eventually appear
on the bottom of the world as well as cells that go left appear on the
right side if they reach the end.
Later, I'd enhance the cells to be able to look at the other side as
currently they're totally blind when going over.

The GridPane is now totally gone and replaced by a Canvas to which I've
painted a proper grid as well as lines that indicate which food each
cells is currently interested in.

Apart from that, the AI received an upgrade! Cells will now check after
each move if the food it was interested in still exists. This somewhat
eliminated so called zombie cells that would spawn and then get beaten
to every food by their parent cells because the cell will simply look
for another food if the first one is eaten by a different cell.

And finally, cells will now attempt to divide, and if it fails, they'll
skip a generation and attempt to divide again. This will repeat until
the cell eventually succeeds in mitosis.