Trees, Leeches and GFX
New cell type: LEECH
It will walk around until it sees a live cell of any type. Once it finds
pray, it'll walk behind the pray and leech it's energy in increments of
5. Like the predator, the Leech will use trails to track down pray.

New cell type: TREE
The tree is unique in two ways: it doesn't move and it can generate
food. It will randomly choose a location in it's FoV and if that tile
doesn't contain food, there is a 30% chance that the Tree will generate
food. Apart from sugar tiles, Trees also feed on corpses.
Energy is gained in increments of 1.

GFX
All cells now have cute 20x20 PNG icons to represent them.